Replace navbar workspace menu with syncfolder menu
