+++ app/stylesheets/shell-window/toolbar-dropdown.less
@@ -108,7 +108,7 @@
   }
 }
 
-.workspace-dropdown-menu {
+.syncfolder-dropdown-menu {
   margin-top: 4px;
   margin-right: 4px;
